# docs/api-spec.yml
openapi: 3.0.0
info:
  title: JobFinder API
  description: |
    REST API for the JobFinder platform, providing core functionality for
    candidate management, job offers, matching, action plans, and
    AI‑driven coaching. The API follows Domain‑Driven Design principles
    and is built on a hybrid Node.js API gateway with a Python agentic core.
  version: 0.1.0
  contact:
    name: JobFinder Development Team
    url: https://github.com/rrangelsegura/JobFinder
  license:
    name: MIT
    url: https://github.com/rrangelsegura/JobFinder/blob/main/LICENSE
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.jobfinder.dev
    description: Production server
...
paths:
  /auth/register:
    post:
      summary: Register a new candidate
      description: |
        Creates a Candidate with a bcrypt-hashed password (cost 12).
        firstName/lastName are not collected here — they're written as
        placeholder values and stay independent of any CV the candidate
        later uploads (see cv-extraction's "Persistence and Embedding of
        Extraction Results"). Sends a single, one-shot reminder email
        prompting a CV upload.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        "201":
          description: Candidate created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Invalid email, password under 8 characters, or email already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
  /auth/login:
    post:
      summary: Log in
      description: |
        On success, creates a Redis-backed session and sets an httpOnly,
        sameSite=lax session cookie (secure in production). Unknown email
        and wrong password return the identical generic error
        (anti-enumeration). Rate-limited per-IP and per-email.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login succeeded; session cookie set.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          description: Invalid email or password (same message for both causes).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "429":
          description: Too many login attempts from this IP or for this email.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
  /auth/session:
    get:
      summary: Check the current session
      description: What ProtectedRoute (frontend) and requireAuth-derived flows gate on.
      tags:
        - Auth
      responses:
        "200":
          description: A valid session exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          description: No valid session cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
...
  /uploads/cv/{jobId}:
    get:
      summary: Get CV extraction job status
      description: |
        Poll for the status of an extraction job enqueued by
        `POST /uploads/cv`. Implements `parse-candidate-cv`'s
        `cv-extraction-status` capability.
      tags:
        - Uploads
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: |
            Job found. `data.status` is `processing`, `completed` (with
            `data.candidate` populated), or `failed` (with `data.error`
            populated).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CVExtractionStatusResponse"
        "404":
          description: No extraction job exists with the given id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadErrorResponse"
  # Additional paths for OCR, coaching, etc., can be added later
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: jobfinder_session
  schemas:
    RegisterResponse:
      type: object
      description: Standard response envelope per docs/backend-standards.md.
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            candidateId:
              type: integer
        agent_trace_id:
          type: string
          format: uuid
        model_used:
          type: string
          nullable: true
    SessionResponse:
      type: object
      description: Standard response envelope per docs/backend-standards.md.
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            candidateId:
              type: integer
            email:
              type: string
              format: email
        agent_trace_id:
          type: string
          format: uuid
        model_used:
          type: string
          nullable: true
    AuthErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        data:
          type: object
          properties:
            error:
              type: string
        agent_trace_id:
          type: string
          format: uuid
        model_used:
          type: string
          nullable: true
...
    CVExtractionStatusResponse:
      type: object
      description: Standard response envelope per docs/backend-standards.md.
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            status:
              type: string
              enum: [processing, completed, failed]
            candidate:
              description: Present only when status is "completed".
              $ref: "#/components/schemas/ExtractedCandidate"
            error:
              type: string
              description: Present only when status is "failed".
        agent_trace_id:
          type: string
          format: uuid
        model_used:
          type: string
          nullable: true
    UploadErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        data:
          type: object
          properties:
            error:
              type: string
        agent_trace_id:
          type: string
          format: uuid
        model_used:
          type: string
          nullable: true
    ExtractedCandidate:
      type: object
      description: |
        The CV Analyst agent's structured extraction output (snake_case, as
        returned by the Python agent and passed through unmodified — distinct
        from the camelCase `Candidate` schema used elsewhere in this API).
      properties:
        personal_info:
          type: object
          properties:
            first_name:
              type: string
            last_name:
              type: string
            email:
              type: string
              format: email
            phone:
              type: string
              nullable: true
            address:
              type: string
              nullable: true
        education:
          type: array
          items:
            type: object
            properties:
              institution:
                type: string
              title:
                type: string
              start_date:
                type: string
                format: date
              end_date:
                type: string
                format: date
                nullable: true
        work_experience:
          type: array
          items:
            type: object
            properties:
              company:
                type: string
              position:
                type: string
              description:
                type: string
                nullable: true
              start_date:
                type: string
                format: date
              end_date:
                type: string
                format: date
                nullable: true
        skills:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [technical, soft]
        languages:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              proficiency:
                type: string
                nullable: true
        certifications:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              issuer:
                type: string
                nullable: true
              issue_date:
                type: string
                format: date
                nullable: true
...